requirements.txt
# Lab 2 NLP - Requirements
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LLM Integration
ollama>=0.1.0
requests>=2.31.0

# MCP Protocol
fastmcp>=0.1.0

# RAG Components
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.0.0

# Data Processing
scipy>=1.11.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Docker Support
docker>=7.0.0

# Jupyter Support (optional)
jupyter>=1.0.0
ipython>=8.0.0
